Move logic into editor-core
