+++ packages/repluggable-core/src/extensionSlot.ts
