+++ packages/repluggable-core/src/index.ts
@@ -1,4 +1,25 @@
 export {
     AnySlotKey,
-    SlotKey
+    SlotKey,
+    AppHost,
+    AppHostAPI,
+    EntryPointOrPackage,
+    Shell
+
 } from './API'
+
+export { createAppHost, makeLazyEntryPoint, mainViewSlotKey, stateSlotKey } from './appHost'
+
+export {AppHostServicesProvider} from './appHostServices'
+
+export { createShellLogger } from './loggers'
+
+export { invokeSlotCallbacks } from './invokeSlotCallbacks'
+
+export { interceptAnyObject } from './interceptAnyObject'
+
+export { monitorAPI } from './monitorAPI'
+
+export { InstalledShellsSelectors, ShellToggleSet } from './installedShellsState'
+
+export { hot } from './hot'
