+++ packages/repluggable-core/src/installedShellsState.ts
