+++ packages/repluggable-core/test/appHost.spec.ts
