+++ packages/repluggable-core/testKit/index.ts
@@ -1,20 +1,15 @@
-import { create, act, ReactTestRenderer, ReactTestInstance, TestRendererOptions } from 'react-test-renderer'
 import _ from 'lodash'
-import React, { ReactElement } from 'react'
-import { createRoot } from 'react-dom/client'
-import { EntryPoint, ObservableState, PrivateShell, ShellBoundaryAspect } from '../src/API'
-import { AnySlotKey, AppHost, AppMainView, createAppHost as _createAppHost, EntryPointOrPackage, Shell, SlotKey } from '../src/index'
-import { ShellRenderer } from '../src/renderSlotComponents'
-import { createShellLogger } from 'repluggable-core'
 
+import { EntryPoint, ObservableState } from '../src/API'
+import { AnySlotKey, AppHost, createAppHost as _createAppHost, SlotKey, EntryPointOrPackage, Shell } from '../src/index'
+
+import { emptyLoggerOptions } from './emptyLoggerOptions'
 
 export { emptyLoggerOptions }
 export { AppHost } from '../src/index'
-export { connectWithShell, connectWithShellAndObserve } from '../src/connectWithShell'
-export { SlotRenderer } from '../src/renderSlotComponents'
 
-export {withThrowOnError, withConsoleErrors} from 'repluggable-core/testKit'
-import { emptyLoggerOptions,  } from 'repluggable-core/testKit'
+export { withConsoleErrors } from './withConsoleErrors'
+export { withThrowOnError } from './withThrowOnError'
 export * from './mockPackage'
 
 export const createAppHost: typeof _createAppHost = (packages, options = emptyLoggerOptions) => {
@@ -119,57 +114,8 @@ export async function createAppHostAndWaitForLoading(packages: EntryPointOrPacka
     return Promise.race([timeoutPromise, loadingPromise]).then(() => appHost)
 }
 
-export const renderHost = (host: AppHost): ReactTestRenderer => {
-    let renderer: ReactTestRenderer | undefined
-    act(() => {
-        renderer = create(<AppMainView host={host} />)
-    })
 
-    return renderer as unknown as ReactTestRenderer
-}
 
-export interface WrappedComponent {
-    testKit: ReturnType<typeof create>
-    host: AppHost
-    parentWrapper: ReactTestInstance | undefined
-}
-
-export const renderDOMInHost = (reactElement: ReactElement<any>, host: AppHost = createAppHost([]), customShell?: Shell) => {
-    const shell = customShell || createShell(host)
-
-    const Component = (
-        <ShellRenderer host={host} shell={shell as PrivateShell} component={<div data-shell-in-host="true">{reactElement}</div>} key="" />
-    )
-
-    const container = document.body.querySelector('div')
-    const root = container && createRoot(container)
-    root?.render(Component)
-}
-
-export const renderInHost = (
-    reactElement: ReactElement<any>,
-    host: AppHost = createAppHost([]),
-    customShell?: Shell,
-    options?: TestRendererOptions
-): WrappedComponent => {
-    const shell = customShell || createShell(host)
-
-    const Component = (
-        <ShellRenderer host={host} shell={shell as PrivateShell} component={<div data-shell-in-host="true">{reactElement}</div>} key="" />
-    )
-    let root: ReactTestRenderer | undefined
-    act(() => {
-        root = create(Component, options)
-    })
-
-    const parentWrapper = root?.root.find(x => x.props['data-shell-in-host'])
-
-    return {
-        testKit: root as unknown as ReactTestRenderer,
-        parentWrapper,
-        host
-    }
-}
 
 type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>
 interface EntryPointOverrides extends Omit<EntryPoint, 'name'> {
@@ -211,58 +157,7 @@ export const addMockShell = (host: AppHost, entryPointOverrides: EntryPointOverr
     return shell
 }
 
-function createShell(host: AppHost): PrivateShell {
-    const entryPoint: EntryPoint = {
-        name: 'test'
-    }
 
-    return {
-        name: entryPoint.name,
-        entryPoint,
-        ...host,
-        declareSlot() {
-            const slot: any = {}
-            return slot
-        },
-        declareCustomSlot() {
-            const slot: any = {}
-            return slot
-        },
-        setLifecycleState: _.noop,
-        setDependencyAPIs: _.noop,
-        canUseAPIs(): boolean {
-            return true
-        },
-        canUseStore(): boolean {
-            return true
-        },
-        wasInitializationCompleted(): boolean {
-            return true
-        },
-        runLateInitializer<T>(initializer: () => T) {
-            return initializer()
-        },
-        contributeAPI<TAPI>(): TAPI {
-            const API: any = {}
-            return API
-        },
-        contributeBoundaryAspect(aspect: ShellBoundaryAspect): void {},
-        getBoundaryAspects(): ShellBoundaryAspect[] {
-            return []
-        },
-        contributeState: _.noop,
-        contributeObservableState: <TState, TSelectors, TAction>() => mockObservable<TSelectors>(undefined as any),
-        contributeMainView: _.noop,
-        flushMemoizedForState: _.noop,
-        memoizeForState: _.identity,
-        memoize: _.identity,
-        clearCache: _.noop,
-        getHostOptions: () => host.options,
-        log: createShellLogger(host, entryPoint),
-        wrapWithShellRenderer: (component: JSX.Element) => component,
-        lazyEvaluator: func => ({ get: func })
-    }
-}
 
 export function mockObservable<T>(value: T): ObservableState<T> {
     return {
@@ -275,8 +170,3 @@ export function mockObservable<T>(value: T): ObservableState<T> {
     }
 }
 
-export function collectAllTexts(instance: ReactTestInstance | undefined) {
-    return (instance
-        ?.findAll(x => x.children?.some(child => typeof child === 'string'))
-        .flatMap(x => x.children.filter(child => typeof child === 'string')) || []) as string[]
-}
