+++ packages/repluggable-core/testKit/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "../dist/testKit/index.js"
+}
