+++ packages/repluggable/src/appMainView.tsx
@@ -2,9 +2,7 @@ import _ from 'lodash'
 import React, { FunctionComponent } from 'react'
 import { connect, Provider } from 'react-redux'
 import { AppHost } from './API'
-import { mainViewSlotKey } from './appHost'
-import { AppHostServicesProvider } from './appHostServices'
-import { InstalledShellsSelectors, ShellToggleSet } from './installedShellsState'
+import { mainViewSlotKey,AppHostServicesProvider, InstalledShellsSelectors, ShellToggleSet } from 'repluggable-core'
 import { SlotRenderer } from './renderSlotComponents'
 import { ShellContext } from './shellContext'
 import { StoreContext } from './storeContext'
